Refactor to include the names
